Add customisable in-game controls (#24)
- Add a pure-CSS expandable `<fieldset>` for changing controls. The
texture options have also been changed to be expandable.
- Each control can be changed to match any number of items or item tags.
Inputs are distinguishable by colour. When tested with Firefox
accessibility tools, they seem to still be distinguishable even with
colourblindness.
- Inputs are using custom form elements, and have `<datalist>`s to show
all item names and tages. Item tags are grabbed from pmmp/BedrockData.
- Add a generalised CachingFetcher class to make caching resource
easier.
- Using a fake material list to translate item names, a list of in-game
controls are added to the pack description if they are different to the
default.