Add in-game translations (kinda)
Pack description formation is VERY MESSY. but it works!!!
Translation of material list and in-game controls in pack description is very hard to do because MCBE translation language names are different from JS language names. Currently, not all languages will be generated. I want to put this in another in-game UI screen where we can translate item names in-game instead of having to translate them for all languages during pack generation. BOO MCBE