Add translation ability to most things (#27)
- Translations can be used for the webpage UI through a JSON file.
- Regular MCBE resource pack `.lang` files can be used for in-game
translations. Some stuff have to be substituted in by the JS but I've
tried to make it as clear as possible...
- I haven't touched some stuff (including pack manifest material list
translation) but most stuff are covered by this. It will be much easier
when I make a dedicated in-game UI screen for Holoprint information
rather than relying on the pack manifest description.
- This is a step towards fully doing #13.