Add Simplified Chinese localization translation (#30)
Translations may contain inaccuracies as not all features have been
tested. When I was testing, I found that some places did not seem to
provide translatable characters. Please fix it.

In order to comply with Chinese word order, some characters/symbols may
be added or removed, so this is not a translation error.