Make layer mode toggleable in-game, many minor improvements (#36)
- Make the layer mode option an in-game control, toggleable using wooden
slabs.
- Since the top layer with all layers below also visible is the same as
the default view, there is no state for that and it skips straight to
the default view.
- Rename "hologram scale" to "ghost block scale"
- Optimise image fetching:
- TextureAtlas.js now avoids fetching the same image URL multiple times
if it appears in multiple image fragments. This significantly reduces
the number of fetch() requests made.
- Make CachingFetchers have 3 fetch() attempts, in case they fail due to
Chromium giving up. This issue was observed when fetching 1600+ images
before the aforementioned optimisation in TextureAtlas.js, and will
probably never happen again due to there only being ~1100 individual
block textures.
- Make the "decrease layer" (logs) in-game control work with the player
as well, not just the armour stand (even though players can crouch with
the regular "change layer" control for the same effect).
- Improve clarity in README.md
- Fix download button sometimes being too small
- Fix `<item-criteria-input>` bugs:
- Fix `<item-criteria-input>`s not being reset when the parent form
reset
- Fix some elements in `<item-criteria-input>` shadow DOMs not
triggering retranslation
  - `<item-criteria-input>`s now remove empty inputs when losing focus
  - `<item-criteria-input>` validation messages are now translatable
  - Fix bug when shift-tabbing in an `<item-criteria-input>`
- Improve spawn animation
- Remove delay at the start of animations if blocks are far away from
the +X+Z corner of the structure
  - Decrease randomness for smaller structures
  - Only animate the first structure in multi-structure packs
- Fix iron bars model