Make block validation particles less annoying
They now have an exclamation mark above the block if they are within 25 blocks of the camera, otherwise the big beam again. This should make it easier to validate layer-by-layer, while still drawing attention to wrong blocks further away. The texture must have lots of padding at the bottom so the center of the particle can stay inside the block space with the exclamation mark floating above.