Finish support for all blocks, add structure updating (#79)
- Using
[pmmp/BedrockBlockUpgradeSchema](https://github.com/pmmp/BedrockBlockUpgradeSchema),
blocks from older Minecraft versions are now automatically updated to
the latest version. This does not do everything perfectly, but it does a
pretty good job at updating everything! I've tested it with some of Mad
Hatter's old test structures from Structura and it seems to handle the
updating pretty well.
- Shulker boxes now show the correct textures and are rotated correctly.
They use the shulker (entity) textures so it's two cubes not one.
- Add jigsaw block rotations
- Add suspicious block brushing progress texture variants
- Add models for beds, heads and skulls
- I wanted to read models/mobs.json to get the geometry, to make it
resource pack compatible, but this will do for 99% of players.
- Beds use a new field in `materialListMappings.json` which adds the
value of a block entity property onto the item aux id for material
lists.
- Heads/Skulls use a new box UV format for cubes, which is found in many
entity models and will be used later when adding entities to holograms.
- Hide "fruit" on jungle leaves
- Fix double slabs
- Update wiki links
- Update to 1.21.60
- Since blocks from older versions are automatically updated to the
latest version (1.21.60), rotations for doors and fence gates had to be
changed.
- Fix unknown variant errors for mycelium, prismarine stairs, and (red)
sandstone stairs
- Bones that have their rotation set to [0, 0, 0] through block state
definitions will now no longer include this redundant property
- Fixed bug with duplicate locators for validation particles in
multi-structure packs
- Allow rotations and texture variants to be defined by block entity
data as well (even though they're in blockStateDefinitions.json...)
- Fix https://github.com/SuperLlama88888/holoprint/issues/77: Resource
packs including files that HoloPrint modifies (namely
armor_stand.entity.json) caused pack generation to fail. This was
because the Response objects created after merging the vanilla file and
the file from the resource pack lost the url property, which was being
used to determine the file path. This caused armor_stand.entity.json to
be read as a Blob. Now, it looks at the file path of the requested
resource rather than the Response url.
- Fix vanilla resource pack stack being duplicated
- Remove gifs from bundles
- Make block translation coverage check run before sample structures
test