Update to 1.21.70, add renaming/retexturing control items, add feature toggles (#96)
- Add all 7 new blocks from the Spring to Life update/drop. The only
problematic ones were leaf litter, which has a unique model, and
wildflowers, which use the pink petals model. However, wildflowers had
two flowers 1px off, so I spent ages trying to find the issue and fix
it... turned out the textures were upside down :(
- Remove redundant definitions for old block (state) names
- Allow block tinting colour to be defined easier
- Control items will now have the name of the control below the item
name in purple. The only side effect is that "..." appears after the
item name if it's in the material list UI. We can't make the material
list rely on alternative translations with the original item names
because they wouldn't exist for languages not supported by HoloPrint.
- Using texture compositing in the item_texture.json texture atlas,
special icons for control textures are now displayed on top of the
regular item icons. Works for most items and blocks with carried
textures. Nextr provided most of the textures, and I tweaked most of
them.
- Add feature toggles for the spawn animation, player controls, the
in-game material list, retexturing control items, and renaming control
items
- Add bug reporting button if pack generation fails
- Add error message for empty files
- Change preview size
- Remove 17k <option> elements
- Datalists for <item-criteria-input>s are now stored in a global
variable, and added/removed on focus/blur.
- Fix nether reactor warning
- Make SimpleLogger a custom element
- Move custom elements on the webpage into the `components` folder
- Use numbers instead of strings for hologram "actions" in the Molang
- Fix pack description translations
- Previously, the pack description for material list and custom control
items would always be in English. Now, it will always be localised.
- When implementing localisation for the first time, I wanted to
separate the concept of a language code as seen in JS navigator.language
and a language code as seen by Minecraft. Now, I'm following the ones
used by Minecraft which makes it easier in the code, hence the rename
from en -> en_US.
- Improve a few JSDoc annotations
- Change default pack name from "HoloPrint.holoprint.mcpack" to
"hologram.holoprint.mcpack"
- Make wiki link in Chinese README go to the Chinese wiki
- Update Supabase SQL for table daily_logs