Fix extremely large structures causing stack overflows (#101)
In Firefox, functions can have at most 500000 parameters; and in Chromium, function parameters are put on a stack, which has a limit of around ~120000. This meant that the spawn animation logic, which removes the initial delay by finding when the first ghost block starts animating, crashed pack generation when there were lots of blocks (I'm talking 814149). This has been fixed with an imperative min function that works limitlessly for arrays.
Also fix the "~ blocks upgraded" info log from not showing up in Chromium browsers.