Fix BigInt JSON serialisation errors (#120)
Block palette entries during palette merging were being added to a JSONSet which didn't handle BigInts correctly. JSONSet has been updated to keep track of all the original values, removing the need for parsing JSON, which means that the only issue that could occur now would be a collision between strings of long numbers and BigInts when in a browser that doesn't support JSON.rawJSON().
JSONMap has also been updated to account for this.