Add mini display size option (#116)
Haven't got up to the optimisation bit yet. Tried doing it with part_visibility in the render controller but that didn't reduce lag. I looked at how Structura does it and it splits the geometry up by layer into separate models and changes which one is being rendered in the render controller.