Fix non-root resource packs folders causing errors (#77)
A misplaced bracket meant that if no manifest.json was found inside a resource pack, an error would be thrown. I suspect this occurred due to a former bug in Chrome on Android in which the full file path would be revealed in File.webkitrelativepath, meaning HoloPrint didn't think there was a manifest.json file.
This has been fixed by determining the root folder which has manifest.json in it, and only looking at the files in that folder. It also makes it easier for me to add full resource pack support soon^tm.