Make full input mode work
I was planning to have a separate UI control for each player control so I could change the item IDs individually, but because the grid_page_indicator creates controls for every subpage in the cycler it would have resulted in n^2 controls total. This just does maths on #collection_index, which required some slight adjustments to the Molang code.